backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
alembic
python-dotenv
groq
langgraph
langchain-core
langchain-community